# Node/UserManagementNodeAPI-JWT/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "User Management API (JWT)",
    "description": "REST API with JWT authentication. **Step 1:** Call `POST /auth/login` with username and password to get a token. **Step 2:** Click **Authorize** and enter `Bearer <your-token>`. **Step 3:** Call protected `/api/users` endpoints.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8084", "description": "Local" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the token from POST /auth/login (without the word Bearer)"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "website": { "type": "string" },
          "address": { "$ref": "#/components/schemas/Address" },
          "company": { "$ref": "#/components/schemas/Company" }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "street": { "type": "string" },
          "suite": { "type": "string" },
          "city": { "type": "string" },
          "zipcode": { "type": "string" },
          "geo": { "$ref": "#/components/schemas/Geo" }
        }
      },
      "Geo": {
        "type": "object",
        "properties": { "lat": { "type": "string" }, "lng": { "type": "string" } }
      },
      "Company": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "catchPhrase": { "type": "string" },
          "bs": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of users", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "username", "email"],
                "properties": {
                  "name": {}, "username": {}, "email": {}, "phone": {}, "website": {},
                  "address": { "$ref": "#/components/schemas/Address" },
                  "company": { "$ref": "#/components/schemas/Company" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  }
}