# Node/UserManagementNodeAPI-MongoDB/server.js
const express = require('express');
const cors = require('cors');
require('dotenv').config();

const { connect } = require('./config/database');
const userRoutes = require('./routes/userRoutes');

const app = express();
const PORT = process.env.PORT || 8084;

// Middleware
app.use(cors());
app.use(express.json());
app.use(express.urlencoded({ extended: true }));

// Routes
app.use('/api/users', userRoutes);

// Health check
app.get('/health', (req, res) => {
    res.status(200).json({ status: 'OK', message: 'User Management API is running' });
});

// 404 handler
app.use((req, res) => {
    res.status(404).json({ message: 'Route not found' });
});

// Error handler
app.use((err, req, res, next) => {
    console.error(err.stack);
    res.status(500).json({ message: 'Internal server error', error: err.message });
});

// Start server after DB connection
async function start() {
    await connect();
    app.listen(PORT, () => {
        console.log(`User Management API server running on http://localhost:${PORT}`);
        console.log(`Health check: http://localhost:${PORT}/health`);
        console.log(`API endpoints: http://localhost:${PORT}/api/users`);
    });
}

start().catch(err => {
    console.error('Failed to start server:', err.message);
    process.exit(1);
});

module.exports = app;